add integration tests for v1beta3
